Added Update/Create Reviews

+++ app/forms/create_review_form.py
@@ -1,7 +1,7 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, IntegerField
+from wtforms import StringField, FloatField
 from wtforms.validators import DataRequired, ValidationError, Length
 
 class CreateReviewForm(FlaskForm):
     review = StringField('Review', validators=[DataRequired(), Length(max=250, message="Review must be less than 250 characters.")])
-    rating = IntegerField('Rating', validators=[DataRequired()])
+    rating = FloatField('Rating', validators=[DataRequired()])
